move provider factories from router to src/providers
